+++ docs/source/changelog.rst
@@ -29,6 +29,8 @@ ChangeLog
 * A profile can now ignore :ref:`settings-ca-default-subject` by setting ``subject`` to ``False``.
 * The ``--pathlen`` and ``--no-pathlen`` parameters for :command:`manage.py init_ca` were changed to
   ``--path-length`` and ``--no-path-length``.
+* Make the Key Usage extension in :command:`manage.py init_ca` configurable.
+* Add support for adding the Inhibit anyPolicy extension when using :command:`manage.py init_ca`.
 * Add support for multiple OCSP responder and CA Issuer entries when creating a certificate authority.
 * Add typehints when installing as wheel.